requirements.txt
git+https://github.com/potatolondon/djangae.git#egg=djangae
git+https://github.com/django/django.git@stable/1.6.x#django
django-secure
django-csp
nose
django-bootstrap3
django-markdown-deux
